requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bayesian-optimization==3.0.0
beautifulsoup4==4.13.4
boto3==1.40.8
botocore==1.40.8
boxmot @ git+https://github.com/mikel-brostrom/boxmot.git@97490df89006be4585935af4ffd8f94ebd9b9402
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
filterpy==1.4.5
fonttools==4.58.4
fsspec==2025.5.1
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.7
lapx==0.5.11.post1
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.9.4
motmetrics==1.4.0
mplcursors==0.6
mpmath==1.3.0
mysql-connector-python==9.4.0
networkx==3.2.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.3.0
polars==1.33.1
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySimpleGUI==5.0.8.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
s3transfer==0.13.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0
xmltodict==1.0.2
yacs==0.1.8
zipp==3.23.0
